[12784] Implement CREATURE_FLAG_EXTRA_ACTIVE
Setting this flag 0x1000 to creature_template.ExtraFlags will result in
the npc being loaded and set to active when the map is created.

In case of continents this will happen on server startup.

Note to DB-Devs: Use this flag with care, it will create more memory and
cpu usage for users.

Note to users: If you have a bottleneck and do not care about i.e. mobs
with huge paths, just remove this flag from the npcs